Move convenient function to more accessible location.
